fix: 实测发现源码写的 u64 但只能写进 32 位

+++ src/lib.rs
@@ -7,7 +7,7 @@ use core::cell::UnsafeCell;
 
 /// SiFive 测试设备。
 #[repr(transparent)]
-pub struct SifiveTestDevice(UnsafeCell<u64>);
+pub struct SifiveTestDevice(UnsafeCell<u32>); // 实测发现源码写的 u64 但只能写进 32 位
 
 const FAIL: u16 = 0x3333;
 const PASS: u16 = 0x5555;
@@ -17,7 +17,7 @@ impl SifiveTestDevice {
     /// 以 code 为错误码，退出进程。
     #[inline]
     pub fn fail(&self, code: u16) -> ! {
-        self.write(FAIL as u64 | (code as u64) << 16)
+        self.write(FAIL as u32 | (code as u32) << 16)
     }
 
     /// 以 0 为错误码，退出进程。
@@ -33,7 +33,7 @@ impl SifiveTestDevice {
     }
 
     #[inline]
-    fn write(&self, bits: u64) -> ! {
+    fn write(&self, bits: u32) -> ! {
         unsafe { self.0.get().write_volatile(bits) };
         unreachable!()
     }
